pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sectorscope"
version = "0.1.0"
description = "テーマ・セクター単位で米国株の騰落率を高速確認できるCLIツール"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "yfinance>=0.2.30",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
sectorscope = "sectorscope.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sectorscope"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
